feat(serializer): Implement serializing methods and implemented __eq__dunder methods
Added a serializer module to handle saving and loading of objects
using both pickle and JSON formats. This enhances the functionality
of the codebase by allowing users to easily persist and retrieve
ElectronicBandStructure and Structure objects.

Includes equality methods for DensityOfStates and ElectronicBandStructure
to facilitate object comparison.

Refactors existing classes to integrate the new serializer methods.