feat(docs): Update documentation structure and content
This commit modifies the documentation configuration by changing the
version import method to use __version__ instead of version. It also
adds new HTML sections to the index.rst file, enhancing the visual
presentation of the documentation with banners and descriptions for
various features of the PyProcar library. These changes aim to improve
user engagement and provide clearer guidance on the library's
capabilities.