refactor(fermisurface): Replace is_band_property method with EBS call
This commit refactors the FermiSurface class by replacing the
is_band_property method with a direct call to the ElectronicBandStructure
instance's is_band_property method. This change simplifies the code
and improves the integration with the ElectronicBandStructure class.

Additionally, the redundant is_band_property method has been removed
to streamline the class functionality.