fix(parser, log_utils): Set default verbosity to 0 and log level to ERROR
This commit updates the Parser class to set the default verbosity level
to 0 during initialization, reducing the amount of output generated by
the parser. Additionally, the logging configuration is modified to set
the log level for both the 'pyprocar' and 'user' loggers to ERROR,
limiting the logs to only error messages. These changes aim to improve
the clarity of the output and focus on critical issues during execution.